requirements.txt
fastapi
uvicorn
ibm-watsonx-ai
python-dotenv
requests
sentence-transformers
chromadb
pypdfium2